FloorMind/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
clip-by-openai>=1.0
sentence-transformers>=2.2.0
flask>=2.3.0
pillow>=10.0.0
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0
requests>=2.31.0
opencv-python>=4.8.0
scikit-learn>=1.3.0